feat: 기존 예상 거리/시간 KM/분 단위로 받던 것 m/ms 단위로 수정

+++ delivery-service/src/main/java/on/logistics/deliveryservice/domain/entity/DeliveryRecord.java
@@ -52,7 +52,7 @@ public class DeliveryRecord extends BaseEntity {
     private UUID endHubId;
 
     @Column(name = "estimated_distance", nullable = false)
-    private double estimatedDistance;
+    private Long estimatedDistance;
 
     @Column(name = "estimated_duration", nullable = false)
     private Long estimatedDuration;
