+++ delivery-service/src/main/java/on/logistics/deliveryservice/infrastructure/client/map/feign/dtos/GetEstimateInfo.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record GetEstimateInfo(Summary summary) {
 
-    public record Summary(int distance, long duration, GetHubRouteInfo.Location start,
+    public record Summary(long distance, long duration, GetHubRouteInfo.Location start,
                           GetHubRouteInfo.Location end, String departureTime, String eta,
                           int tollFare) {
